+++ app/views/shared/_email_footer.html.erb
@@ -2,7 +2,7 @@
   <table class='email-footer' align='center'>
     <tr>
       <td class='email-footer-text'>
-        Please direct any questions to <%= mail_to(Convection.config.contact_email) %> or call
+        Please direct any questions to <%= mail_to(Convection.config.contact_email_address) %> or call
         <%= link_to Convection.config.contact_phone_number, "tel:#{Convection.config.contact_phone_number}" %>.
       </td>
     </tr>
