+++ app/views/user_mailer/submission_receipt.html.erb
@@ -59,7 +59,7 @@
             <table class='medium-width-table response-info-table' align='center'>
               <tr>
                 <td class='response-info'>
-                  Our current response time ranges 2-3 weeks. Please direct any questions to <%= mail_to(Convection.config.contact_email) %>
+                  Our current response time ranges 2-3 weeks. Please direct any questions to <%= mail_to(Convection.config.contact_email_address) %>
                   or call <%= link_to Convection.config.contact_phone_number, "tel:#{Convection.config.contact_phone_number}" %>.
                 </td>
               </tr>
