+++ config/initializers/_config.rb
@@ -7,9 +7,10 @@ Convection.config = OpenStruct.new(
   admin_email_address: ENV['ADMIN_EMAIL_ADDRESS'] || 'consign@artsy.net',
   artsy_url: ENV['ARTSY_URL'] || 'https://staging.artsy.net',
   cloudfront_url: ENV['CLOUDFRONT_URL'],
-  contact_email: ENV['CONTACT_EMAIL'] || 'specialist@artsy.net',
+  contact_email_address: ENV['CONTACT_EMAIL'] || 'specialist@artsy.net',
   contact_phone_number: ENV['CONTACT_PHONE_NUMBER'] || '+1 (646) 712-8154',
   convection_url: ENV['CONVECTION_URL'] || 'https://convection-staging.artsy.net',
+  debug_email_address: ENV['DEBUG_EMAIL_ADDRESS'] || 'sarah@artsymail.com',
   gemini_account_key: ENV['GEMINI_ACCOUNT_KEY'] || 'convection-staging',
   gemini_app: ENV['GEMINI_APP'] || 'https://media.artsy.net',
   gemini_s3_key: ENV['GEMINI_S3_KEY'] || 'replace-me',
