+++ spec/views/admin/submissions/show.html.erb_spec.rb
@@ -18,7 +18,8 @@ describe 'admin/submissions/show.html.erb', type: :feature do
         edition_size: 100,
         edition_number: '23a',
         category: 'Painting',
-        user_id: 'userid')
+        user_id: 'userid',
+        state: 'submitted')
 
       stub_jwt_header('userid')
       page.visit "/admin/submissions/#{@submission.id}"
